requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.20.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0
tqdm>=4.64.0
jupyter>=1.0.0
datasets>=2.10.0 